Show service online status
Adds a column to the services table that displays the online
status of a service using traffic lights, much like the
respective column in the servers table.
This change should make it easier to reason about why
certain jobs may not able to be dispatched.

Notably, an offline service does not necessarily mean that
there is an error or even a problem. Therefore this patch
does not add a notification about services
being offline.

Requires https://github.com/opencast/opencast/pull/6632 to function.