Make textFilter persist across tabs
Same as #980, but for the textfilter, aka the search field.

When typing something into the search bar, then switching to
a different tab, then switching back to the original tab,
the search bar is empty. With this patch, the search bar will
instead remember what was originally written in it.