Click-to-filter in series table
Allows users to click on organizers, contributors and
creation dates in the series table to filter the table by the
value that was clicked on, similar to how it works in the events table.

Requires backend changes to work. Should not break anything
without the backend changes, the buttons will simply not be
responsive.