Slightly improve indentation in `DetailsMetadataTab` (#1324)
* Fix indentation/nesting

* Remove a redundant conditions and subsequently improve indentation

For `someArray: SomeType[]`, `!!someArray``is never `false`.
Also there is no need to check the `length`; if it is `0`,
`Array.prototype.map` will just do nothing.

The rest is indentation.

* Slightly improve indentation in `DetailsMetadataTab`

This is mostly just indentation and breaking lines differently,
but in order to facilitate that, I rephrased some conditions.

* Restore parens around `return`

As per reviewer request